requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
cryptography==39.0.2
Deprecated==1.2.13
Django==4.1.7
djangorestframework==3.14.0
gcloud==0.18.3
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
numpy==1.24.2
oauth2client==4.1.3
opencv-python==4.7.0.72
Pillow==9.5.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==4.0.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
wrapt==1.15.0
